Prevent Doctrine2 YAML from writing out M2M tables.
Amends writeTable() so that it does not write out M2M tables.  This
problem appears to affect most writeTable functions and a better
solution would be push the logic of whether to write (i.e. external
tables and m2m tables) should be in the base table, leaving the actual
writing to the children.  If there is interest, I am happy to provide a
fix along those lines.